@@ -23,7 +23,7 @@
     "@types/node": "^22.9.0",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
-    "eslint": "^9.14.0",
+    "eslint": "^9.15.0",
     "eslint-config-next": "^15.0.3",
     "next": "^14.0.4",
     "react": "^18.3.1",
