@@ -56,8 +56,9 @@
     "@babel/preset-env": "^7.26.0",
     "@babel/preset-react": "^7.25.9",
     "@babel/preset-typescript": "^7.26.0",
-    "@eslint/compat": "^1.2.2",
-    "@types/node": "^20.10.5",
+    "@eslint/compat": "^1.2.3",
+    "@redux-saga/core": "^1.3.0",
+    "@types/node": "^22.9.0",
     "@types/react": "^18.3.12",
     "@types/react-dom": "^18.3.1",
     "@types/sanitize-html": "^2.13.0",
@@ -66,16 +67,15 @@
     "@typescript-eslint/eslint-plugin": "^8.14.0",
     "@typescript-eslint/parser": "^8.14.0",
     "@vitejs/plugin-react": "^4.3.3",
-    "eslint": "^9.14.0",
+    "eslint": "^9.15.0",
     "eslint-plugin-babel": "^5.3.1",
-
     "eslint-plugin-react": "^7.37.2",
     "eslint-plugin-react-hooks": "^5.0.0",
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "eslint-plugin-unused-imports": "^4.1.4",
     "globals": "^15.12.0",
     "lerna": "^8.1.9",
-    "nx": "^20.1.0",
+    "nx": "^20.1.2",
     "react-spring": "^9.7.4",
     "sanitize-html": "^2.13.1",
     "semver": "^7.6.3",
@@ -89,9 +89,8 @@
     "@sentry/electron": "^4.15.1",
     "@sentry/react": "^7.91.0",
     "@sentry/tracing": "^7.91.0",
-    "chokidar": "^3.5.3",
+    "chokidar": "^4.0.1",
     "cross-env": "^7.0.3",
-    "hw-fingerprint": "^2.1.0",
     "keytar": "^7.9.0",
     "prettier": "^3.3.3",
     "pretty-bytes": "^6.1.1",
