@@ -1,18 +1,18 @@
 /* eslint-disable no-sync */
 /* eslint-disable @typescript-eslint/no-var-requires */
-const asar = require('asar');
-const path = require('path');
-const fs = require('fs');
-const os = require('os');
+import asar from 'asar';
+import path from 'path';
+import fs from 'fs';
+import os from 'os';
 
 const architectures = ['ia32', 'x64', 'armv7l', 'arm64', 'universal'];
 
-exports.default = async function afterPack(context) {
+export default async function afterPack(context) {
 	const arch = architectures[context.arch];
 	const platform = context.packager.platform.nodeName;
 	const tempDirPath = path.join(os.tmpdir(), Date.now().toString());
 	const asarPath = generateAsarPath(platform, context.appOutDir);
-	const nativeKeytarDir = path.join(__dirname, '..', '..', '..', 'build', 'native-components', 'keytar');
+	const nativeKeytarDir = path.join(import.meta.dirname, '..', '..', '..', 'build', 'native-components', 'keytar');
 	const nativeKeytarPath = path.join(nativeKeytarDir, generateKeytarFilename(arch, platform));
 	const asarKeytarPath = path.join(tempDirPath, 'keytar.node');
 
