@@ -1,8 +1,8 @@
 import { IpcEvent } from '@beak/common/ipc/ipc';
+import { BrowserWindow, IpcMainInvokeEvent } from 'electron';
 import path from 'path';
 
 import { getProjectFilePathWindowMapping, getProjectIdFromWindowId } from './fs-shared';
-import { BrowserWindow, IpcMainInvokeEvent } from 'electron';
 
 export function getProjectFolder(event: IpcEvent) {
 	const projectFilePath = getProjectFilePathWindowMapping(event);
