@@ -1,7 +1,7 @@
 import { EditorPreferences } from '@beak/common/types/beak-hub';
 import { ipcFsService } from '@beak/ui/lib/ipc';
+import { call, select } from '@redux-saga/core/effects';
 import path from 'path-browserify';
-import { call, select } from 'redux-saga/effects';
 
 import { ApplicationState } from '../..';
 
