@@ -1,8 +1,8 @@
 import { RequestPreference } from '@beak/common/types/beak-hub';
 import { ipcFsService } from '@beak/ui/lib/ipc';
+import { call, select } from '@redux-saga/core/effects';
 import { PayloadAction } from '@reduxjs/toolkit';
 import path from 'path-browserify';
-import { call, select } from 'redux-saga/effects';
 
 import { ApplicationState } from '../..';
 import { RequestPreferencePayload } from '../types';
