@@ -2,8 +2,8 @@ import { changeTab } from '@beak/ui/features/tabs/store/actions';
 import { duplicateRequestNode } from '@beak/ui/lib/beak-project/request';
 import { ipcDialogService } from '@beak/ui/lib/ipc';
 import type { Nodes } from '@getbeak/types/nodes';
+import { call, delay, put, race, select, take } from '@redux-saga/core/effects';
 import { PayloadAction } from '@reduxjs/toolkit';
-import { call, delay, put, race, select, take } from 'redux-saga/effects';
 
 import { ApplicationState } from '../..';
 import { ActionTypes, DuplicateRequestPayload } from '../types';
