@@ -4,8 +4,8 @@ import { ShowMessageBoxRes } from '@beak/common/ipc/dialog';
 import { attemptReconciliation } from '@beak/ui/features/tabs/store/actions';
 import { removeVariableGroup } from '@beak/ui/lib/beak-variable-group';
 import { ipcDialogService } from '@beak/ui/lib/ipc';
+import { call, put } from '@redux-saga/core/effects';
 import { PayloadAction } from '@reduxjs/toolkit';
-import { call, put } from 'redux-saga/effects';
 
 import actions from '../actions';
 import { RemoveVariableGroupFromDiskPayload } from '../types';
